update user repo

+++ src/main/java/com/my_task/service/user/UserResponse.java
@@ -6,8 +6,8 @@ import lombok.Builder;
 
 @Builder
 public record UserResponse(String id, String firstName, String lastName, String email, String avatarUrl,
-		Boolean enabled, long numOfTodo,
-		long numOfCompleted, long totalTasks) {
+		Boolean enabled, int numOfTodo,
+		int numOfCompleted, int totalTasks) {
 	public static UserResponse form(User user) {
 		return UserResponse.builder().id(user.getId()).email(user.getEmail()).firstName(user.getFirstName())
 				.avatarUrl(user.getAvatarUrl())
